Enhance the sync command
Previously the sync used to compute the sync plans manually (fetch the list of users from source directory, compare with what should be in the destination directories). In some cases, the computation is not that easy. Now we use the more advanced multi-services user which can map some users between services more accurately.