Enhance the dependency management
Now the project compiles with a simple swift build on macOS, but with a ton of LDAP deprecation warnings.
If you want to avoid the warnings you have to compile using the build_for_macos script or generate the xcodeproj with the generate_xcodeproj script. This is due to OpenLDAP not providing a pkg-config file and thus preventing us to force SPM to use Homebrew’s OpenLDAP…